+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# 0.4.3
+- Added SkillTreeNode#previousNodeIds and SkillTreeNode#getPreviousNodes().
+- Made LeveledSkillTreeNode extend SkillTreeNode.
 # 0.4.2
 - Renamed StatProperty#statPropertyType to StatProperty#type.
 - Renamed typedef RelicType to RelicTypeId, and added RelicType class.